Rename `failure` package to `error`
