+++ sbt-github/src/main/scala/com/alejandrohdezma/sbt/github/error/NotFound.scala
@@ -1,4 +1,4 @@
-package com.alejandrohdezma.sbt.github.failure
+package com.alejandrohdezma.sbt.github.error
 
 import scala.util.control.NoStackTrace
