+++ sbt-github/src/test/scala/com/alejandrohdezma/sbt/github/json/DecoderSpec.scala
@@ -19,7 +19,7 @@ package com.alejandrohdezma.sbt.github.json
 import java.time.ZoneOffset.UTC
 import java.time.ZonedDateTime
 
-import com.alejandrohdezma.sbt.github.failure._
+import com.alejandrohdezma.sbt.github.error._
 import com.alejandrohdezma.sbt.github.json.Decoder.Failure._
 import com.alejandrohdezma.sbt.github.syntax.json._
 import org.specs2.mutable.Specification
