+++ sbt-github/src/test/scala/com/alejandrohdezma/sbt/github/syntax/JsonSyntaxSpec.scala
@@ -16,7 +16,7 @@
 
 package com.alejandrohdezma.sbt.github.syntax
 
-import com.alejandrohdezma.sbt.github.failure._
+import com.alejandrohdezma.sbt.github.error._
 import com.alejandrohdezma.sbt.github.json.Decoder.Failure.NotABoolean
 import com.alejandrohdezma.sbt.github.json.Json
 import com.alejandrohdezma.sbt.github.json.Json.Failures.{InvalidPath, NotAJSONObject}
